Wire entire review command on the new architecture
CU6 migrates the entire review command from the legacy 1,513-line
cmd/entire/cli/review.go into the cmd/entire/cli/review/ subpackage,
routing user-facing entry points through CU1-CU5b's new abstractions:
env-handshake, AgentReviewer/Sink/Run, ReviewerTemplate,
ComposeReviewPrompt, scope detection.

Surface preserved (matches PR #993 + #1009):
  entire review                          — main command
  entire review --edit                   — re-open picker
  entire review --agent <name>           — override agent
  entire review attach <session-id>      — post-hoc attach
  entire review attach {--force,--agent,--skills}

--track-only is intentionally NOT included (dropped by #1009).

Architecture:
  - review/cmd.go: NewCommand() with Deps struct; runReview() dispatches
    to launchable agents (claude/codex/gemini → AgentReviewer.Run() with
    DumpSink) or non-launchable (cursor/opencode/factoryai-droid →
    marker_fallback.go which writes PendingReviewMarker + prints
    guidance — single shared file, no per-agent fallback files).
  - review/picker.go: pickConfig + first-run setup migrated from old code.
  - review/attach_test.go: smoke test for the migrated subcommand.
  - cli/review_bridge.go + review_helpers.go: small bridge files in the
    cli package holding functions that can't move into review/ due to
    import cycles (review → checkpoint → codex → review;
    review → claudecode/codex/geminicli → review). Bridge wires
    cli-package implementations into review.NewCommand's Deps struct.

Deletes cmd/entire/cli/review.go and review_test.go (1,513 + ~600
lines). All still-used functions migrated; the rest were already
replaced by CU1-CU5b. Per-feature tests now live alongside their
implementations under review/.

.golangci.yaml: add review/types.AgentReviewer to ireturn allow-list
(was Process only); the new bridge legitimately returns the interface.
